Add files via upload
The purpose of this project is to store the daily work status of the employees.This project involves external hyperlinks, dynamic charts, pivot table, filtering, and specialized functions such as date. Also it deals with absolute & relative cell references, restrictions, drop down list, and so on.The output of this project provides a crystal clear pictorial representation of individual employee work reports, monthly reports, project wise reports, and yearly reports. These reports are helpful for employee assessment, project effort calculation, and future project estimations and so on